pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentapi"
version = "0.1.0"
description = "Lightweight framework for building agentic AI backends with FastAPI"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
keywords = ["agents", "llm", "fastapi", "ai", "tool-calling"]
authors = [
  { name = "AgentAPI" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Framework :: FastAPI",
  "Topic :: Software Development :: Libraries",
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1"
]

[project.urls]
Homepage = "https://github.com/prajwalsuryawanshi/agentapi"
Repository = "https://github.com/prajwalsuryawanshi/agentapi"
Issues = "https://github.com/prajwalsuryawanshi/agentapi/issues"
Documentation = "https://github.com/prajwalsuryawanshi/agentapi#readme"

[project.scripts]
agentapi = "agentapi.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
agentapi = ["assets/*.png"]

[tool.setuptools.packages.find]
include = ["agentapi*"]

[tool.setuptools.exclude-package-data]
"*" = ["__pycache__/*", "*.pyc"]
